Heatmap/Contour Plot for Assessing Power of the CQA PPQ Plan Using General Multiplier.

The function for plotting the heatmap to evaluate the PPQ plan based on the specification test, given lower and upper specification limits.

PPQ_ctplot(attr.name, attr.unit, Llim, Ulim, mu, sigma, n, n.batch, k, test.point)

Arguments

attr.name
(optional) user-defined attribute name for PPQ assessment

attr.unit
(optional) user-defined attribute unit

Llim
lower specification limit

Ulim
upper specification limit

mu
hypothetical mean of the attribute

sigma
hypothetical standard deviation of the attribute

n
sample size (number of locations) per batch

n.batch
number of batches for passing PPQ during validation

k
general multiplier for constructing the specific interval

test.point
(optional) actual process data points for testing whether the processes pass PPQ

Value

Heatmap (or Contour Plot) for PPQ Assessment.

References

Burdick, R. K., LeBlond, D. J., Pfahler, L. B., Quiroz, J., Sidor, L., Vukovinsky, K., & Zhang, L. (2017). Statistical Applications for Chemistry, Manufacturing and Controls (CMC) in the Pharmaceutical Industry. Springer.

See also

Examples

if (FALSE) {
mu <- seq(1.6,3.4,0.05)
sigma <- seq(0.05,0.8,0.01)
PPQ_ctplot(attr.name = "Total Protein", attr.unit = "mg/mL", Llim=1.5, Ulim=3.5,
mu = mu, sigma = sigma, k=2.373)

## Example verifying simulation resutls in the textbook page 249
mu <- seq(95, 105, 0.1)
sigma <- seq(0.2, 5, 0.1)
PPQ_ctplot(attr.name = "Composite Assay", attr.unit = "%LC", Llim=95, Ulim=105,
mu = mu, sigma = sigma, k=2.373)
mu <- seq(90, 110, 0.5)
PPQ_ctplot(attr.name = "Composite Assay", attr.unit = "%LC", Llim=90, Ulim=110,
mu = mu, sigma = sigma, k=2.373)

mu <- seq(95,105,0.1)
sigma <- seq(0.1,2.5,0.1)
PPQ_ctplot(attr.name = "Sterile Concentration Assay", attr.unit = "%", Llim=95, Ulim=105,
mu = mu, sigma = sigma, k=2.373)
test <- data.frame(mean=c(97,98.3,102.5), sd=c(0.55, 1.5, 1.2))
PPQ_ctplot(attr.name = "Sterile Concentration Assay", attr.unit = "%", Llim=95, Ulim=105,
mu = mu, sigma = sigma, k=2.373, test.point=test)
}